apps/api/pyproject.toml
[project]
name = "sa-call-analyzer"
version = "0.1.0"
description = "AI agent that analyzes Solution Architect performance on sales calls using Command of the Message framework"
readme = "../../README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "anthropic>=0.18.1",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "crewai>=0.11.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "langchain-litellm>=0.1.0",
    "litellm>=1.0.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "openinference-instrumentation-litellm>=0.1.0",
    "arize-otel>=0.1.0",
    "arize-phoenix-client>=1.0.0",
    "requests>=2.31.0",
    "python-pptx>=0.6.21",
    "python-docx>=1.1.0",
    "rapidfuzz>=3.0.0",
    "google-cloud-bigquery>=3.0.0",
    "openinference-instrumentation-openai>=0.1.0",
    "langgraph>=0.2.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-threading>=0.45b0",
]

[project.optional-dependencies]
# chroma / langchain-chroma — RAG demo code (arize_demo_traces/use_cases/rag.py). Note: CrewAI still pulls chromadb (~hundreds of MB).
chroma = [
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
]
aws-bedrock = [
    "langchain-aws>=0.2.0",
]
hypothesis = [
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.0.0",
]
full = [
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
    "langchain-aws>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
# Tells Vercel where to load the FastAPI `app` (see deploy docs — main.py is not a default entry name).
app = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "*.py",
    "frontend/*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: timing-based tests that may be flaky in CI",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
